refactor(agent): split session/turn.rs (1908 lines) into turn/ directory
Sub-modules: core (turn lifecycle), tools (execution + delegation),
context (history + prompt), session_io (transcript persistence).
