+++ src/openhuman/agent/harness/session/turn/core.rs
@@ -1,0 +1,615 @@
+//! Core turn execution: the main `turn()` method and `inject_agent_experience_context()`.
+
+use super::super::types::Agent;
+use super::super::transcript;
+use super::super::turn_engine_adapter::{AgentCheckpoint, AgentObserver, AgentToolSource};
+use super::{integration_announcement_note, normalize_tool_call};
+use crate::openhuman::agent::harness;
+use crate::openhuman::agent::hooks::{self, TurnContext};
+use crate::openhuman::agent::memory_loader::collect_recall_citations;
+use crate::openhuman::agent::progress::AgentProgress;
+use crate::openhuman::agent_experience::{
+    prepend_experience_block, render_experience_hits, AgentExperienceStore, ExperienceQuery,
+};
+use crate::openhuman::inference::provider::{ChatMessage, ConversationMessage};
+use crate::openhuman::memory::MemoryCategory;
+use crate::openhuman::util::truncate_with_ellipsis;
+
+use anyhow::Result;
+use std::hash::{Hash, Hasher};
+use std::sync::Arc;
+
+impl Agent {
+    /// Executes a single interaction "turn" with the agent.
+    ///
+    /// This function is the primary driver of the agent's behavior. It manages the
+    /// end-to-end lifecycle of a user request:
+    ///
+    /// 1. **Initialization**: Resumes from a session transcript if this is a new turn
+    ///    to preserve KV-cache stability.
+    /// 2. **Prompt Construction**: Builds the system prompt (only on the first turn)
+    ///    incorporating learned context and tool instructions.
+    /// 3. **Context Injection**: Enriches the user message with relevant memories
+    ///    fetched via the [`MemoryLoader`].
+    /// 4. **Execution Loop**: Enters a loop (up to `max_tool_iterations`) where it:
+    ///    - Manages the context window (reduction/summarization).
+    ///    - Calls the LLM provider.
+    ///    - Parses and executes tool calls.
+    ///    - Accumulates results into history.
+    /// 5. **Synthesis**: Returns the final assistant response after all tools have
+    ///    finished or the iteration budget is exhausted.
+    /// 6. **Background Tasks**: Triggers episodic memory indexing and facts
+    ///    extraction asynchronously.
+    pub async fn turn(&mut self, user_message: &str) -> Result<String> {
+        let turn_started = std::time::Instant::now();
+        self.emit_progress(AgentProgress::TurnStarted).await;
+        log::info!("[agent] turn started — awaiting user message processing");
+        log::info!(
+            "[agent_loop] turn start message_chars={} history_len={} max_tool_iterations={}",
+            user_message.chars().count(),
+            self.history.len(),
+            self.config.max_tool_iterations
+        );
+        self.ensure_composio_integrations_listener();
+        // ── Session transcript resume ─────────────────────────────────
+        // On a fresh session (empty history), look for a previous
+        // transcript to pre-populate the exact provider messages for
+        // KV cache prefix reuse.
+        if self.history.is_empty() && self.cached_transcript_messages.is_none() {
+            self.try_load_session_transcript();
+        }
+
+        if self.history.is_empty() {
+            // Learned context is only baked into the system prompt on the
+            // very first turn — once the history is non-empty we reuse the
+            // stored prompt verbatim to preserve the KV-cache prefix the
+            // inference backend has already tokenised. Fetching it later
+            // would just burn memory-store reads on data we throw away.
+            if !self.connected_integrations_initialized {
+                self.fetch_connected_integrations().await;
+                // Sessions born without a cached Composio view still need
+                // a one-shot delegation-surface reconcile before the system
+                // prompt is frozen. The shared-Arc failure path returns
+                // `false`, but on turn 1 the Arc should still be uniquely
+                // owned; a `false` return here indicates a programmer error
+                // and the warn-level log inside the helper already surfaces
+                // it, so we keep the existing best-effort contract.
+                let _ = self.refresh_delegation_tools();
+            }
+            let learned = self.fetch_learned_context().await;
+            let rendered_prompt = self.build_system_prompt(learned)?;
+            log::info!("[agent] system prompt built — initialising conversation history");
+            log::info!(
+                "[agent_loop] system prompt built chars={}",
+                rendered_prompt.chars().count()
+            );
+            // User-file injection (PROFILE.md, MEMORY.md) puts
+            // potentially-sensitive content (LinkedIn scrape output,
+            // archivist-curated memories) into the system prompt. Avoid
+            // leaking that to debug logs — log a length + content hash
+            // instead. Narrow specialists (both flags off) keep the
+            // full-body log so prompt-engineering iteration on
+            // tools/safety sections stays easy.
+            if self.omit_profile && self.omit_memory_md {
+                log::debug!("[agent_loop] system prompt body:\n{}", rendered_prompt);
+            } else {
+                let mut hasher = std::collections::hash_map::DefaultHasher::new();
+                rendered_prompt.hash(&mut hasher);
+                log::debug!(
+                    "[agent_loop] system prompt body redacted (contains PROFILE/MEMORY): chars={} hash={:016x}",
+                    rendered_prompt.chars().count(),
+                    hasher.finish()
+                );
+            }
+            self.history
+                .push(ConversationMessage::Chat(ChatMessage::system(
+                    rendered_prompt,
+                )));
+            // Seed the per-turn mid-session refresh baseline with the
+            // hash of whatever Composio actually returned just now.
+            // Subsequent turns short-circuit unless this hash changes.
+            self.last_seen_integrations_hash =
+                crate::openhuman::composio::connected_set_hash(&self.connected_integrations);
+            // Seed the announced set with the startup connected toolkits so
+            // only genuinely-new mid-session connects get announced later.
+            self.announced_integrations = self
+                .connected_integrations
+                .iter()
+                .map(|i| i.toolkit.clone())
+                .collect();
+        } else {
+            // Deliberately do NOT rebuild the system prompt on subsequent
+            // turns. The rendered prompt is the KV-cache prefix the inference
+            // backend has already tokenised; replacing its bytes (even
+            // cosmetically) forces the backend to re-prefill from scratch.
+            //
+            // Dynamic turn-to-turn context (memory recall, learned snippets)
+            // rides on the user message via `memory_loader.load_context()`
+            // — that's where the caller should inject anything that varies
+            // between turns.
+            //
+            // *** Mid-session schema-only refresh ***
+            //
+            // The system prompt stays frozen, but the function-calling
+            // schema (the `tools` field in the provider request) is sent
+            // fresh on every API call — it's not part of the KV-cache
+            // prefix. So we *can* react to Composio connect/disconnect
+            // events mid-session by re-synthesising the `delegate_<toolkit>`
+            // surface on `self.tools` / `self.tool_specs` and letting
+            // the next provider call carry the new schema. KV cache stays
+            // intact; the system prompt's `## Connected Integrations`
+            // block goes mildly stale until the next session, but the
+            // schema is the source of truth the model actually routes
+            // against.
+            //
+            // The signal we react to is the process-wide
+            // [`crate::openhuman::composio::INTEGRATIONS_CACHE`], kept
+            // current by (a) the desktop UI's 5 s
+            // `composio_list_connections` poll, (b) the post-OAuth
+            // `ComposioConnectionCreatedSubscriber` invalidation, and
+            // (c) the 60 s TTL fallback. We read it via the read-only
+            // [`crate::openhuman::composio::cached_active_integrations`]
+            // helper — never trigger a backend fetch ourselves, never
+            // block on a writer.
+            // Session agents built through `from_config_*` carry their
+            // runtime `Config` snapshot directly, so this read avoids the
+            // old `Config::load_or_init()` round-trip on every turn.
+            //
+            let _ = self.refresh_delegation_tools_from_cached_integrations("turn-boundary");
+            // Cache empty/expired or config unavailable => no signal.
+            // We leave the current tool surface alone and pick up any
+            // real change on the next turn after the UI's 5 s poll has
+            // repopulated [`INTEGRATIONS_CACHE`].
+
+            log::trace!(
+                "[agent_loop] system prompt reused (history_len={}) — KV cache prefix preserved",
+                self.history.len()
+            );
+        }
+
+        if self.auto_save {
+            let _ = self
+                .memory
+                .store(
+                    "",
+                    "user_msg",
+                    user_message,
+                    MemoryCategory::Conversation,
+                    None,
+                )
+                .await;
+        }
+
+        log::info!("[agent] loading memory context for user message");
+        const MEMORY_CITATION_LIMIT: usize = 5;
+        const MEMORY_CITATION_MIN_RELEVANCE: f64 = 0.4;
+        match collect_recall_citations(
+            self.memory.as_ref(),
+            user_message,
+            MEMORY_CITATION_LIMIT,
+            MEMORY_CITATION_MIN_RELEVANCE,
+        )
+        .await
+        {
+            Ok(citations) => {
+                log::debug!(
+                    "[agent_loop] memory citations collected count={}",
+                    citations.len()
+                );
+                self.last_turn_citations = citations;
+            }
+            Err(err) => {
+                log::warn!("[agent_loop] memory citation collection failed: {err}");
+                self.last_turn_citations.clear();
+            }
+        }
+        let context = self
+            .memory_loader
+            .load_context(self.memory.as_ref(), user_message)
+            .await
+            .unwrap_or_default();
+
+        // ── Phase 3 STM preemptive recall ────────────────────────────
+        // On the very first turn only, assemble a bounded cross-thread
+        // context block from the FTS5 episodic arm (keyword match) and the
+        let mut context = context;
+
+        // ── Lane B: situational preferences (every turn) ─────────────────────
+        // Recall topic-scoped preferences semantically relevant to THIS message
+        // (model-aware embeddings, gated by vector similarity) and inject them
+        // under a banner. Runs every turn — unlike the first-turn-gated tree/STM
+        // blocks above — because the query changes per message; it rides the
+        // per-turn context that's prepended to the user message (no KV-cache
+        // cost). An unrelated message clears the similarity gate to nothing, so
+        // no block is injected.
+        {
+            let situational =
+                crate::openhuman::memory::preferences::recall_situational_preferences(
+                    &self.memory,
+                    user_message,
+                )
+                .await;
+            if !situational.is_empty() {
+                log::info!(
+                    "[pref_recall] situational block injected: {} item(s)",
+                    situational.len()
+                );
+                context.push_str("## Relevant preferences for this message\n\n");
+                for pref in &situational {
+                    context.push_str("- ");
+                    context.push_str(pref.trim());
+                    context.push('\n');
+                }
+                context.push('\n');
+            } else {
+                log::debug!("[pref_recall] no situational preference relevant to this message");
+            }
+        }
+
+        let enriched = if context.is_empty() {
+            log::info!("[agent] no memory context found — using raw user message");
+            self.last_memory_context = None;
+            user_message.to_string()
+        } else {
+            log::info!(
+                "[agent] memory context loaded — enriching user message context_chars={}",
+                context.chars().count()
+            );
+            self.last_memory_context = Some(context.clone());
+            format!("{context}{user_message}")
+        };
+
+        let enriched = self
+            .inject_agent_experience_context(user_message, enriched)
+            .await;
+
+        // ── SKILL.md body injection (#781) ───────────────────────────
+        // Match installed SKILL.md skills against the user message and
+        // prepend their bodies ahead of the memory-context block so the
+        // LLM sees them at the top of the user turn. See the module
+        // docs on [`crate::openhuman::workflows::inject`] for the matching
+        // heuristic and size cap rationale.
+        let enriched = {
+            use crate::openhuman::workflows::inject;
+            let matches = inject::match_workflows(&self.skills, user_message);
+            if matches.is_empty() {
+                log::debug!(
+                    "[skills:inject] no skill matches for user message (skill_catalog_len={})",
+                    self.skills.len()
+                );
+                enriched
+            } else {
+                let injection = inject::render_injection(
+                    &matches,
+                    inject::DEFAULT_MAX_INJECTION_BYTES,
+                    |skill| skill.read_body(),
+                );
+                let matched_count = injection.decisions.iter().filter(|d| d.matched).count();
+                log::info!(
+                    "[skills:inject] summary candidates={} matched={} injected_bytes={} truncated_any={}",
+                    injection.decisions.len(),
+                    matched_count,
+                    injection.injected_bytes,
+                    injection.truncated
+                );
+                if injection.rendered.is_empty() {
+                    enriched
+                } else {
+                    format!("{}\n{}", injection.rendered, enriched)
+                }
+            }
+        };
+
+        // Consume any one-shot mid-session connect announcement parked by
+        // `refresh_delegation_tools_from_cached_integrations`. It rides on the
+        // user turn (NOT a system message — `trim_history` hoists system
+        // messages to the front and would bust the KV-cache prefix) and
+        // `.take()` clears it so it fires exactly once.
+        let pending_slugs = std::mem::take(&mut self.pending_integration_announcement);
+        let enriched = match integration_announcement_note(&pending_slugs) {
+            Some(note) => format!("{note}\n\n{enriched}"),
+            None => enriched,
+        };
+
+        self.history
+            .push(ConversationMessage::Chat(ChatMessage::user(enriched)));
+
+        // Pin the main agent to its configured model for the lifetime of
+        // the session. Per-turn classification used to run here, but it
+        // would flip `effective_model` mid-conversation (e.g. reasoning →
+        // coding based on a single keyword). Every flip invalidates the
+        // backend's KV cache namespace for this session, costing full
+        // re-prefill on the very next turn. The main agent's job is to
+        // decide *which sub-agent* to spawn — that routing lives in the
+        // model prompt, not in the Rust-side classifier. Sub-agents pick
+        // their own tier via `ModelSpec::Hint(...)` in their definition.
+        let effective_model = self.model_name.clone();
+        log::info!(
+            "[agent_loop] model pinned model={} (per-turn classification disabled for KV cache stability)",
+            effective_model
+        );
+
+        // Snapshot the parent's runtime once per turn so any
+        // `spawn_subagent` invocation that fires inside this turn can
+        // read it via the PARENT_CONTEXT task-local. We override the
+        // model field with the post-classification effective model.
+        let mut parent_context = self.build_parent_execution_context();
+        parent_context.model_name = effective_model.clone();
+
+        // Bump the session-memory turn counter. Used later by
+        // `should_extract_session_memory` to decide whether to spawn a
+        // background archivist fork at end-of-turn.
+        self.context.tick_turn();
+
+        let turn_body = async {
+            // Capture everything the engine seams need as locals/clones *before*
+            // the observer takes `&mut self`, so the borrow checker is happy:
+            // the tool source + parser + checkpoint hold clones disjoint from
+            // the `Agent`, and the observer alone borrows it mutably.
+            let dispatcher = self.tool_dispatcher.clone();
+            let provider = self.provider.clone();
+            let provider_name = self.event_channel().to_string();
+            let temperature = self.temperature;
+            let max_iterations = self.config.max_tool_iterations;
+            // Source multimodal limits from the session's runtime config when
+            // present so [IMAGE:…] / [FILE:…] markers in user messages are
+            // resolved with the operator-configured caps (max files, max size,
+            // max extracted text). Without this, agents fall back to the
+            // crate-default caps and `MultimodalFileConfig::default()`
+            // disables file expansion entirely.
+            let multimodal = self
+                .integration_runtime_config
+                .as_ref()
+                .map(|c| c.multimodal.clone())
+                .unwrap_or_default();
+            let multimodal_files = self
+                .integration_runtime_config
+                .as_ref()
+                .map(|c| c.multimodal_files.clone())
+                .unwrap_or_default();
+            let artifact_store = Some(
+                crate::openhuman::agent::harness::tool_result_artifacts::ToolResultArtifactStore::new(
+                    self.action_dir.clone(),
+                    self.session_key.clone(),
+                ),
+            );
+            let mut tool_source = AgentToolSource {
+                tools: self.tools.clone(),
+                visible_tool_names: self.visible_tool_names.clone(),
+                tool_policy_session: self.tool_policy_session.clone(),
+                tool_policy: self.tool_policy.clone(),
+                payload_summarizer: self.payload_summarizer.clone(),
+                event_session_id: self.event_session_id().to_string(),
+                event_channel: self.event_channel().to_string(),
+                agent_definition_id: self.agent_definition_id.clone(),
+                prefer_markdown: self.context.prefer_markdown_tool_output(),
+                budget_bytes: self.context.tool_result_budget_bytes(),
+                artifact_store: artifact_store.clone(),
+                should_send_specs: self.tool_dispatcher.should_send_tool_specs(),
+                advertised_specs: self.visible_tool_specs.as_ref().clone(),
+                records: Vec::new(),
+            };
+            let progress = super::super::super::engine::TurnProgress::new(self.on_progress.clone());
+            let parser = super::super::super::engine::DispatcherParser {
+                dispatcher: dispatcher.as_ref(),
+            };
+            let checkpoint = AgentCheckpoint {
+                provider: self.provider.clone(),
+                dispatcher: self.tool_dispatcher.clone(),
+                model: effective_model.clone(),
+                temperature,
+                on_progress: self.on_progress.clone(),
+                user_message: user_message.to_string(),
+                max_iterations,
+            };
+            let turn_run_queue = self.run_queue.clone();
+            let cached_prefix = self.cached_transcript_messages.take();
+            let mut observer = AgentObserver {
+                agent: self,
+                artifact_store,
+                effective_model: effective_model.clone(),
+                cumulative_input: 0,
+                cumulative_output: 0,
+                cumulative_cached: 0,
+                cumulative_charged: 0.0,
+                last_turn_usage: None,
+                cached_prefix,
+                pending_results: Vec::new(),
+                did_push_final: false,
+            };
+            let mut buf: Vec<ChatMessage> = Vec::new();
+
+            // Box-pin the parent agent's engine call so its ~600-line
+            // generator state lives on the heap. Tools that delegate to
+            // sub-agents (orchestrator → researcher / personality /
+            // archetype / skill) recurse back into another
+            // `run_turn_engine` via `run_subagent`; without the box,
+            // both engines' state machines pile up on the same tokio
+            // worker stack and overflow the 2 MiB default. The inner
+            // boxes inside `run_typed_mode` aren't reached if the
+            // overflow happens during the parent's poll on the way in
+            // — verified against the `chat-harness-subagent` Playwright
+            // lane crash on PR #3151.
+            let outcome = Box::pin(super::super::super::engine::run_turn_engine(
+                provider.as_ref(),
+                &mut buf,
+                &mut tool_source,
+                &progress,
+                &mut observer,
+                &checkpoint,
+                &parser,
+                &provider_name,
+                &effective_model,
+                temperature,
+                true, // silent — the channel/UI renders via progress + the return value
+                &multimodal,
+                &multimodal_files,
+                max_iterations,
+                None, // the web bridge streams via on_progress deltas, not on_delta
+                &[],
+                turn_run_queue,
+            ))
+            .await?;
+
+            // Pull the observer's accounting out, then drop it to release the
+            // `&mut self` borrow so the epilogue can use `self`.
+            let did_push_final = observer.did_push_final;
+            let cumulative_input = observer.cumulative_input;
+            let cumulative_output = observer.cumulative_output;
+            let cumulative_cached = observer.cumulative_cached;
+            let cumulative_charged = observer.cumulative_charged;
+            let last_turn_usage = observer.last_turn_usage.take();
+            drop(observer);
+            let records = std::mem::take(&mut tool_source.records);
+
+            self.context.record_tool_calls(records.len());
+
+            // For a clean final response the observer already pushed the
+            // assistant message + persisted. For a max-iteration checkpoint or
+            // circuit-breaker halt the engine returned the text without pushing
+            // it, so finish the history + transcript here (mirrors the old
+            // final/max-iter branches).
+            if !did_push_final {
+                self.history
+                    .push(ConversationMessage::Chat(ChatMessage::assistant(
+                        outcome.text.clone(),
+                    )));
+                self.trim_history();
+                // Note: the engine already emits `TurnCompleted` on the
+                // checkpoint exit (and every other terminal path), so we don't
+                // re-emit it here — doing so would double-fire for the UI.
+                let messages = self.tool_dispatcher.to_provider_messages(&self.history);
+                self.persist_session_transcript(
+                    &messages,
+                    cumulative_input,
+                    cumulative_output,
+                    cumulative_cached,
+                    cumulative_charged,
+                    last_turn_usage.as_ref(),
+                );
+            }
+
+            // Auto-save a short memory of the final reply (not on a capped turn,
+            // matching the prior behavior).
+            if self.auto_save && !outcome.hit_cap {
+                let summary = truncate_with_ellipsis(&outcome.text, 100);
+                let _ = self
+                    .memory
+                    .store("", "assistant_resp", &summary, MemoryCategory::Daily, None)
+                    .await;
+            }
+
+            // Fire post-turn hooks (non-blocking).
+            if !self.post_turn_hooks.is_empty() {
+                let ctx = TurnContext {
+                    user_message: user_message.to_string(),
+                    assistant_response: outcome.text.clone(),
+                    tool_calls: records,
+                    turn_duration_ms: turn_started.elapsed().as_millis() as u64,
+                    session_id: Some(self.event_session_id.clone())
+                        .filter(|session_id| !session_id.trim().is_empty()),
+                    agent_id: Some(self.agent_definition_id.clone())
+                        .filter(|agent_id| !agent_id.trim().is_empty()),
+                    entrypoint: Some(self.event_channel.clone())
+                        .filter(|entrypoint| !entrypoint.trim().is_empty()),
+                    iteration_count: outcome.iterations as usize,
+                };
+                hooks::fire_hooks(&self.post_turn_hooks, ctx);
+            }
+
+            Ok(outcome.text)
+        }; // end of `turn_body` async block
+
+        // Run the turn body inside the parent-execution-context scope so
+        // that any `spawn_subagent` tool call fired during the loop can
+        // read the parent's provider, tools, model, and workspace via
+        // the PARENT_CONTEXT task-local.
+        let result = harness::with_parent_context(parent_context, turn_body).await;
+
+        // Session transcript persistence lives INSIDE the turn body —
+        // one write per provider response, fired right after the
+        // response lands (see the tool-call and terminal branches in
+        // `turn_body`). A crash during tool execution no longer drops
+        // the assistant's reply because it was already flushed to
+        // disk before tool dispatch started. No outer-loop save is
+        // needed here.
+
+        // ── Session-memory extraction (stage 5) ───────────────────────
+        //
+        // If the pipeline's deltas have crossed all three thresholds
+        // (token growth, tool calls, turn count), spawn a *background*
+        // archivist sub-agent that will distil durable facts into the
+        // workspace MEMORY.md file via the `update_memory_md` tool.
+        //
+        // The spawn is fire-and-forget: the main turn returns the
+        // user-visible response immediately, and the archivist runs
+        // asynchronously on the `agentic` tier. We optimistically mark
+        // the extraction complete right away — if it actually fails,
+        // we'll just retry on the next threshold window (a few turns
+        // later), which is the right amount of retry behaviour for a
+        // librarian task that's idempotent across reruns.
+        if result.is_ok() && self.context.should_extract_session_memory() {
+            self.spawn_session_memory_extraction().await;
+            // Sibling pipeline (#1399): heuristic transcript ingestion
+            // turns the just-written transcript into durable
+            // conversational memory + reflections so a brand-new chat
+            // can recover continuity. Background-only, never blocks the
+            // user-facing turn return.
+            self.spawn_transcript_ingestion();
+        }
+
+        result
+    }
+
+    pub(super) async fn inject_agent_experience_context(
+        &self,
+        user_message: &str,
+        enriched: String,
+    ) -> String {
+        const MAX_EXPERIENCE_HITS: usize = 3;
+        const MAX_EXPERIENCE_BLOCK_BYTES: usize = 2048;
+
+        if !self.learning_enabled {
+            return enriched;
+        }
+
+        let tools = self
+            .visible_tool_specs
+            .iter()
+            .map(|spec| spec.name.clone())
+            .collect();
+        let store = AgentExperienceStore::new(self.memory.clone());
+        let query = ExperienceQuery {
+            query: user_message.to_string(),
+            tools,
+            tags: Vec::new(),
+            agent_id: Some(self.agent_definition_id.clone()).filter(|id| !id.trim().is_empty()),
+            entrypoint: Some(self.event_channel.clone())
+                .filter(|entrypoint| !entrypoint.trim().is_empty()),
+            max_hits: MAX_EXPERIENCE_HITS,
+        };
+
+        match store.retrieve(query).await {
+            Ok(hits) => {
+                let matched_hits: Vec<_> = hits
+                    .into_iter()
+                    .filter(|hit| !hit.match_reasons.is_empty())
+                    .collect();
+                let block = render_experience_hits(&matched_hits, MAX_EXPERIENCE_BLOCK_BYTES);
+                if block.is_empty() {
+                    return enriched;
+                }
+                log::debug!(
+                    "[agent-experience] injected {} experience hit(s) bytes={}",
+                    matched_hits.len(),
+                    block.len()
+                );
+                prepend_experience_block(&enriched, &block)
+            }
+            Err(err) => {
+                log::warn!("[agent-experience] retrieval failed (non-fatal): {err}");
+                enriched
+            }
+        }
+    }
+}
